feat: Added structure to the project
